Port ts-aot to build and produce ELF executables on Linux (x86_64)
Cross-platform changes behind #ifdef guards - Windows behavior unchanged.

Build system:
- CMake: conditional vcpkg triplet, platform-specific LLVM/LLD components
- Compiler CMakeLists: LLD ELF linking, -Wl,--start-group for circular deps
- Runtime CMakeLists: pthread/dl linking, portable vcpkg paths
- ICU data stub: version-aware via U_ICU_VERSION_MAJOR_NUM macro

Compiler:
- Driver.cpp: Linux path layout (single-config, 3 levels vs 4 to project root)
- LinkerDriver.cpp: ELF linking with CRT files, system libs, full path resolution
- ExtensionLoader.cpp: /proc/self/exe for executable path
- ICU dat filename made version-aware throughout

Runtime:
- TsGC.cpp: pthread-based stack bounds for conservative GC scanning
- StackMap.cpp: read ELF sections from /proc/self/exe (not memory-mapped)
- TsError.cpp: backtrace() + __cxa_demangle for Linux stack traces
- Core.cpp: POSIX signal handlers, ICU version-aware dat filename
- EventLoop.cpp, TsPromise.cpp, Primitives.cpp: removed unguarded windows.h

Portability fixes:
- Lambda-to-void* casts: unary + or explicit function pointer (GCC)
- SCHED_RR/SCHED_NONE renamed to avoid Linux sched.h macro conflict
- int64_t vs long long in extern "C" signatures
- __declspec(noinline) -> TS_NOINLINE macro
- InliningPass.h: explicit constructor for GCC aggregate init compatibility

+++ extensions/node/os/include/TsOs.h
@@ -7,6 +7,8 @@
 #ifndef TS_OS_EXT_H
 #define TS_OS_EXT_H
 
+#include <stdint.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -29,9 +31,9 @@ void* ts_os_get_eol();
 void* ts_os_get_devnull();
 
 // Memory functions
-long long ts_os_totalmem();
-long long ts_os_freemem();
-long long ts_os_availableParallelism();
+int64_t ts_os_totalmem();
+int64_t ts_os_freemem();
+int64_t ts_os_availableParallelism();
 
 // Other info functions
 double ts_os_uptime();
@@ -42,8 +44,8 @@ void* ts_os_networkInterfaces();
 void* ts_os_userInfo();
 
 // Priority functions
-long long ts_os_getPriority(long long pid);
-void ts_os_setPriority(long long pid, long long priority);
+int64_t ts_os_getPriority(int64_t pid);
+void ts_os_setPriority(int64_t pid, int64_t priority);
 
 // Constants (objects)
 void* ts_os_get_constants();
